feat(userManagement): Implement user table columns correctly
